Done
Sheldon, Leonard, Wolowitz, Koothrappally and Penny are in line for the "double cola". As soon as a person drinks such a cola, he duplicate and both stand at the end of the queue for more drinks cup. Write a program that will display queue status depending on how many glasses of cola issued a device with a wonderful drink.