Done
Implement a web interface to the Group class. On the screen browser should display a list of methods that you can do with the group. When clicking on the method link it must be executed. The result of the work, in the form of html page should be returned to the user.